src/ai/requirements.txt
numpy
pandas
google-generativeai
python-dotenv 